MacQuest_Serverside/requirement.txt
Django==2.1.4
djangorestframework==3.9.0
pytz==2018.7
Whoosh==2.7.4
